Implement sum reduction in money
The commit includes the addition of a 'Sum' class that enables the addition of two monetary amounts. Modifications have been made to the 'Money' class, 'Expression' interface, 'Bank' class, and 'MoneyTest' to allow testing of the new functionality. New methods, 'getAmount' and 'reduce', have been added to improve efficiency and coverage of use cases.